+++ src/Types/Enums/Type.hpp
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <cstddef>
 #include <cstdint>
 
 namespace pokesim::dex {
@@ -28,5 +27,5 @@ enum class Type : std::uint8_t {
   TYPE_TOTAL
 };
 
-static constexpr std::size_t TOTAL_TYPE_COUNT = (std::size_t)Type::TYPE_TOTAL - 1U;
+static constexpr std::uint8_t TOTAL_TYPE_COUNT = (std::uint8_t)Type::TYPE_TOTAL - 1U;
 }  // namespace pokesim::dex
